+++ clios/exceptions.py
@@ -2,26 +2,23 @@ from typing import TypedDict
 
 from pydantic import ValidationError
 
-
-class TokenError(Exception):
-    def __init__(self, token: str = "", pos: int = 0, msg: str = "") -> None:
-        self.pos = pos
-        self.token = token
-        if msg:
-            super().__init__(msg)
-        else:
-            super().__init__()
+from .cli.tokenizer import Token
 
 
 class CliosError(Exception):
     class CliosErrorCtx(TypedDict, total=False):
-        token_index: int
-        unchainable_token_index: int
-        expected_num_args: int
-        arg_key: str
-        arg_index: int
+        token: Token
+        unchainable_token: Token
         validation_error: ValidationError
 
     def __init__(self, msg: str = "", ctx: CliosErrorCtx = {}) -> None:
         self.ctx = ctx
         super().__init__(msg)
+
+
+class ParserError(CliosError):
+    pass
+
+
+class OperatorError(CliosError):
+    pass
