+++ tests/test_parse_tokens.py
@@ -2,59 +2,68 @@
 import pytest
 from pydantic import ValidationError
 
-from clios import ParserError
-from clios.ast_builder import ErrorType, _Empty
-from clios.tokenizer import LeftBracketToken
-from clios.tokenizer import OperatorToken as Ot
-from clios.tokenizer import StringToken as Ft
+from clios.cli.ast_builder import _Empty
+from clios.cli.tokenizer import LeftBracketToken
+from clios.cli.tokenizer import OperatorToken as Ot
+from clios.cli.tokenizer import StringToken as Ft
+from clios.exceptions import ParserError
 
 invalids = [
     [
         [Ft("f2"), Ot("op_not_found")],
-        ParserError(ErrorType.OPERATOR_NOT_FOUND, ctx={"token_index": 0}),
+        ParserError("Operator `f2` not found!", ctx={"token_index": 0}),
     ],
     [
         [Ot("op_1i"), Ot("op_not_found")],
-        ParserError(ErrorType.OPERATOR_NOT_FOUND, ctx={"token_index": 1}),
+        ParserError("Operator `op_not_found` not found!", ctx={"token_index": 1}),
     ],
     [
         [Ot("op_1i"), LeftBracketToken()],
-        ParserError(ErrorType.SYNTAX_NOT_SUPPORTED, ctx={"token_index": 1}),
+        ParserError("This syntax is not supported yet!", ctx={"token_index": 1}),
     ],
     [
         [Ot("op_not_found")],
-        ParserError(ErrorType.OPERATOR_NOT_FOUND, ctx={"token_index": 0}),
+        ParserError("Operator `op_not_found` not found!", ctx={"token_index": 0}),
     ],
     [
         [Ot("op_1o_noroot")],
-        ParserError(ErrorType.UNSUPPORTED_ROOT_OPERATOR, ctx={"token_index": 0}),
+        ParserError(
+            "Operator `op_1o_noroot` cannot be used as root operator!",
+            ctx={"token_index": 0},
+        ),
     ],
     [
         [
             Ot("op_1i,1"),
         ],
         ParserError(
-            ErrorType.TOO_MANY_ARGS, ctx={"token_index": 0, "expected_num_args": 0}
+            "Too many arguments: expected 0, got 1",
+            ctx={"token_index": 0},
         ),
     ],
     [
         [
             Ot("op_1i,a=1"),
         ],
-        ParserError(ErrorType.UNEXPECTED_KWD, ctx={"token_index": 0, "arg_key": "a"}),
+        ParserError(
+            "Unexpected keyword argument: `a`!", ctx={"token_index": 0, "arg_key": "a"}
+        ),
     ],
     [
         [
             Ot("op_1i1k,ip=1,a=1"),
         ],
-        ParserError(ErrorType.UNEXPECTED_KWD, ctx={"token_index": 0, "arg_key": "a"}),
+        ParserError(
+            "Unexpected keyword argument: `a`!", ctx={"token_index": 0, "arg_key": "a"}
+        ),
     ],
     [
         [
             Ot("op_1i1k"),
         ],
         ParserError(
-            ErrorType.MISSING_REQUIRED_KWD, ctx={"token_index": 0, "arg_key": "ip"}
+            "Missing required keyword argument: `ip`!",
+            ctx={"token_index": 0, "arg_key": "ip"},
         ),
     ],
     [
@@ -62,7 +71,8 @@ invalids = [
             Ot("op_1i1p"),
         ],
         ParserError(
-            ErrorType.TOO_FEW_ARGS, ctx={"token_index": 0, "expected_num_args": 1}
+            "Missing arguments: expected atleast 1, got 0",
+            ctx={"token_index": 0},
         ),
     ],
     [
@@ -70,27 +80,28 @@ invalids = [
             Ot("op_1i1p,1,2"),
         ],
         ParserError(
-            ErrorType.TOO_MANY_ARGS, ctx={"token_index": 0, "expected_num_args": 1}
+            "Too many arguments: expected 1, got 2",
+            ctx={"token_index": 0},
         ),
     ],
     [
         [Ot("op_vi"), Ot("op_1o"), Ot("op_1o"), Ot("op_1o"), Ot("op_")],
         ParserError(
-            ErrorType.CHAIN_TYPE_ERROR,
+            "These operators cannot be chained together!",
             ctx={"token_index": 0, "unchainable_token_index": 4},
         ),
     ],
     [
         [Ot("op_2i"), Ot("op_1o"), Ot("op_1i1o"), Ot("op_")],
         ParserError(
-            ErrorType.CHAIN_TYPE_ERROR,
+            "These operators cannot be chained together!",
             ctx={"token_index": 2, "unchainable_token_index": 3},
         ),
     ],
     [
         [Ot("op_2i"), Ot("op_1o"), Ot("op_")],
         ParserError(
-            ErrorType.CHAIN_TYPE_ERROR,
+            "These operators cannot be chained together!",
             ctx={"token_index": 0, "unchainable_token_index": 2},
         ),
     ],
@@ -104,29 +115,29 @@ invalids = [
             Ot("op_"),
         ],
         ParserError(
-            ErrorType.CHAIN_TYPE_ERROR,
+            "These operators cannot be chained together!",
             ctx={"token_index": 0, "unchainable_token_index": 5},
         ),
     ],
     [
         [Ot("op_1i1o")],
-        ParserError(ErrorType.MISSING_OUTPUT, ctx={"token_index": 0}),
+        ParserError("Missing output(s)!", ctx={"token_index": 0}),
     ],
     [
         [Ot("op_1i1o"), Ot("op_1o"), Ot("op_1o")],
-        ParserError(ErrorType.MISSING_OUTPUT_FILE, ctx={"token_index": 2}),
+        ParserError("Output file path must be a string", ctx={"token_index": 2}),
     ],
     [
         [Ot("op_1i"), Ot("op_1o"), Ot("op_1o")],
-        ParserError(ErrorType.TOO_MANY_INPUTS, ctx={"token_index": 2}),
+        ParserError("Got too many inputs!", ctx={"token_index": 2}),
     ],
     [
         [Ot("op_vi")],
-        ParserError(ErrorType.TOO_FEW_INPUTS, ctx={"token_index": 0}),
+        ParserError("Missing inputs!", ctx={"token_index": 0}),
     ],
     [
         [Ot("op_1i")],
-        ParserError(ErrorType.MISSING_INPUTS, ctx={"token_index": 0}),
+        ParserError("Missing inputs!", ctx={"token_index": 0}),
     ],
 ]
 
@@ -134,7 +145,7 @@ build_validation_error = [
     [
         [Ot("op_1p,a")],
         {
-            "error_type": ErrorType.ARG_VALIDATION_ERROR,
+            "error_type": "Data validation error!",
             "token_index": 0,
             "arg_index": 0,
         },
@@ -142,7 +153,7 @@ build_validation_error = [
     [
         [Ot("op_1k,ip=a")],
         {
-            "error_type": ErrorType.ARG_VALIDATION_ERROR,
+            "error_type": "Data validation error!",
             "token_index": 0,
             "arg_key": "ip",
         },
@@ -150,7 +161,7 @@ build_validation_error = [
     [
         [Ot("op_1i"), Ft("input")],
         {
-            "error_type": ErrorType.INPUT_VALIDATION_ERROR,
+            "error_type": "Data validation error!",
             "token_index": 1,
         },
     ],
@@ -169,7 +180,7 @@ def test_invalid(parser, input, expected):
 def test_validation_error(parser, input, expected):
     with pytest.raises(ParserError) as e:
         parser.parse_tokens(input)
-    assert str(e.value) == expected["error_type"].value
+    assert str(e.value) == expected["error_type"]
     assert e.value.ctx["token_index"] == expected["token_index"]
     assert isinstance(e.value.ctx["validation_error"], ValidationError)
     if "arg_index" in expected:
